+++ changelog.md
@@ -1,5 +1,21 @@
 ﻿# TILER2 Changelog
 
+**2.2.2**
+
+- Made FakeInventory more compatible with BetterUI and other item sorting mods.
+
+**2.2.1**
+
+- Updated BetterUI hooks for v1.5.7.
+
+**2.2.0**
+
+- Migrated and publicized mod compat classes from ClassicItems. Now provides public hooks for BetterUI, ItemStats, and ShareSuite.
+
+**2.1.3**
+
+- Fixed items dropping while disabled when R2API.ItemDropAPI was loaded by another mod.
+
 **2.1.2**
 
 - Fixed inability of FakeInventory to prevent item steal.